requirements.txt
tensorflow
pandas
dvc
notebook
numpy
matplotlib
seaborn
python-box
pyYAML
tqdm #progress bar
ensure
joblib
types-pyYAML
scipy
Flask
Flask-cors
#-e .
